finish sign up api
