+++ lib/core/constants/app_constants.dart
@@ -1,4 +1,5 @@
 class AppConstants {
   static const String baseUrl = 'https://ecommerce.routemisr.com/api/v1/';
   static const String loginPath = 'auth/signin';
+  static const String signUpPath = 'auth/signup';
 }
